better sorting of page results
